1. Початок. Базова навігація 1.01 Навігація 1.1 Навігація закінчена. Виклик за номером 1.2 Internet, Map 1.2.1 Щось зробив 1.3 Вдстань до груп 1.3.1 1.4 Зробив визов через API з тестового сайта 1.5 Під'єднав Room 1.5.1 Пофіксив Room 1.6 Добавиви на вкладку Home дати 1.6.1 Date Store 1.7 Заповнив Home Screen 1.7.1 1.8 Змнив графік на сайті та в App 1.8.1 GroupListScreen підсвчування активних груп 1.9 HomeScreen preferences 1.9.1 Schedule:Lit<String> 1.8.2 обавив щоденні роздуми 1.8.3 Додав іконкі до телефонів 1.8.4. Data Store refactoring 1.9. ReadScreen, FontFamily 1.9.1 GroupLit add Filter 2. Під'єднались до https://gso.aa.kiev.ua/ 2.1 GroupScreen use UIAction and UIErrorState 2.2 Включив підтримку озвучування в HomeScreen(TTS) 2.2.1 Refactoring HomeScreen 2.2.2 Add Prayers Ukr 3. DI Hit 3.1.1 Refactoring DI 3.2 Refactoring 3.3 Add update check for App 3.4 Add update App from GitHub Release 3.4.1 Add Info on HomeScreen 3.4.2 Add Info on GroupList, Group 4.1.1 Collapsing Card on HomeScreen 4.1.2 Add jvmTarget = JvmTarget.JVM_11 in bild.gradle 5. Creating a new Branch for Navigation 3 5.1 Navigation 3. Start 5.1.1 Navigation 3 works 5.1.2 changed class Prayer.kt and add PrayerText 5.1.3 deleting installed APK 5.1.4 Add transaction Animation 5.1.5 Merge branches 5.1.6 rebuild ReedScreen 5.1.7 minor refactoring 5.1.8 Refactiring: repository, GroupList

+++ app/src/main/java/com/alexyach/compose/groupsaa/data/repository/GroupsRepositoryImpl.kt
@@ -3,61 +3,79 @@ package com.alexyach.compose.groupsaa.data.repository
 import android.util.Log
 import com.alexyach.compose.groupsaa.data.db.GroupDao
 import com.alexyach.compose.groupsaa.data.db.GroupEntity
+import com.alexyach.compose.groupsaa.data.db.toGroup
 import com.alexyach.compose.groupsaa.data.model.toGroup
 import com.alexyach.compose.groupsaa.data.network.ApiService
 import com.alexyach.compose.groupsaa.domain.model.Group
 import com.alexyach.compose.groupsaa.domain.model.toGroupEntity
 import com.alexyach.compose.groupsaa.domain.repository.IRepository
+import com.alexyach.compose.groupsaa.presentation.grouplist.LoadingListGroup
+import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.Flow
+import kotlinx.coroutines.flow.MutableStateFlow
+import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.catch
+import kotlinx.coroutines.flow.first
 import kotlinx.coroutines.flow.flow
+import kotlinx.coroutines.flow.flowOn
+import java.io.IOException
 import javax.inject.Inject
 import javax.inject.Singleton
 
 @Singleton
 class GroupsRepositoryImpl @Inject constructor(
-    val groupDao: GroupDao,
-    val apiService: ApiService
+    private val groupDao: GroupDao,
+    private val apiService: ApiService,
+    private val networkHelper: NetworkHelper // check Internet
 ) : IRepository {
 
 
-    override fun getAllGroupList() = flow {
+    /* New Code */
+    override fun getAllGroupList(): Flow<Pair<List<Group>, Boolean>> = flow {
 
-        Log.d("Logs", "retrofitImpl() getAllGroupList Start")
+        /* Сначала всегда показываем кэш из Room (быстро) */
+//        emit(Pair<List<Group>, Boolean>(downloadFromRoom(), false))
 
-        /* TEST */
-//        throw Exception("This is an error message.")
-//        emit(getListGroupTest())
-        /* END TEST */
+        /* Если есть интернет — обновляем из сети */
+        if (networkHelper.isOnline()) {
+            try {
+                /* TEST */ //throw IOException()
 
-        val groupsDTO = apiService.getGroupList()
-        emit( groupsDTO.map { it.toGroup() }) // to viewModel
+                val groupsFromInternet = apiService.getGroupList().map { it.toGroup() }
+                emit(Pair<List<Group>, Boolean>(groupsFromInternet, true))
 
+                /* Обновить Room*/
+                updateRoomGroups(groupsFromInternet)
 
-    }.catch {
-        Log.d("Logs", "retrofitImpl() Error : $it")
-    }
+            } catch (e: IOException) {
+                Log.i("Logs", "Не вдалося оновити дані")
+                Log.i("Logs", "Помилка $e")
+                emit(Pair<List<Group>, Boolean>(downloadFromRoom(), false))
+            }
 
-
-    override fun getAllFromRoom(): Flow<List<GroupEntity>> {
-        Log.d("Logs", "GroupsRepositoryImpl getAllFromRoom")
-
-        return groupDao.getAll()
-    }
-
-    override suspend fun saveToRoom(groupsDto: List<Group>) {
-        Log.d("Logs", "GroupsRepositoryImpl saveToRoom")
-
-        groupsDto.forEach { item ->
-            groupDao.insertGroup(item.toGroupEntity())
+        } else {
+            emit(Pair<List<Group>, Boolean>(downloadFromRoom(), false))
         }
+    }.flowOn(Dispatchers.IO)
 
+    private suspend fun downloadFromRoom(): List<Group> {
+
+        return groupDao.getAll().first().map { it.toGroup() }
     }
 
-    override suspend fun deleteAllFromRoom() {
-        groupDao.deleteAll()
+    private suspend fun updateRoomGroups(listFromInternet: List<Group>) {
 
-        Log.d("Logs", "GroupsRepositoryImpl deleteAllFromRoom")
+        val listFromRoom: List<Group> = downloadFromRoom()
+
+        if (listFromInternet != listFromRoom) {
+            if (listFromRoom.isNotEmpty()) {
+                groupDao.deleteAll()
+            }
+
+            listFromInternet.forEach { group ->
+                groupDao.insertGroup(group.toGroupEntity())
+            }
+        }
     }
